JVMCBC-108: Check if endpoint is active, not only writable.
Motivation
----------
The current code checked, before writing into a channel, if it
isWritable(). As it turns out, this is not what it says, since
it only checks if the underlying buffer can be written immediately.

Modifications
-------------
The changeset makes sure that the code calls isActive() as well to
make sure its both open, active and writable.

Result
------
Avoiding situations where data is written down to a channel which
is not active anymore, avoiding broken pipes and similar.

Change-Id: I1621d2f68985023ef955383c0257fdf89374251c
Reviewed-on: http://review.couchbase.org/44801
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>
Reviewed-on: http://review.couchbase.org/44908
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Tested-by: Simon Baslé <simon@couchbase.com>