JVMCBC-110: Upgrade Netty to 4.0.25
Motivation
----------
The current codebase is on 4.0.25, the last two bugfix releases
bring both general speed improvements and lots of bugfixes in the SSL
area.

It also contains faster UTF-8 string encoding capabilities which we
can utilize in the future.

Modifications
-------------
Upgrade Netty to 4.0.25.

Result
------
Newer stable netty release with better performance and bugfixes.

Change-Id: If71955bf9ff7cfb044e69e9a619e4f010fad3cb1
Reviewed-on: http://review.couchbase.org/44809
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>