JVMCBC-114: Do not release K/V content prematurely.
Motivation
----------
Since some operations can be retried on "not my vbucket", its content
cannot be released until we are sure that it is not needed anymore.

This is only true for content() which arrives from an external scope,
so the observe command needs to be spared (since it creates the content
in scope).

Modifications
-------------
Retaining content in the k/v handler where needed, and then the decode
code makes sure to release them properly (if not RETRY).

Unit tests have been added to cover the appropriate commands.

Result
------
Correct behavior during rebalance scenarios when ops need to be retried.

Change-Id: I7cb6789ca3b7dbe2c78a6c6027a6580d4277461e
Reviewed-on: http://review.couchbase.org/44840
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>
Reviewed-on: http://review.couchbase.org/44914
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Tested-by: Simon Baslé <simon@couchbase.com>