JVMCBC-115: Adding a minimum reconnect delay to the Endpoint.
Motivation
----------
Currently the code starts with very very low reconnect delays (2, 4, 8,..)
ms. When a node is removed or failed over, this leads to lots of connect
errors.

A better approach is to provide a longer initial min reconnect delay which
is then gradually increased.

Modifications
-------------
This changeset sets a minimum delay of 128ms, which is still very quick. It does
this for a few attempts and then increases up to 4s as before.

In future versions, once the delay is merged we can think about making it configurable
through the environment completely.

Result
------
More graceful reconnect delays in the lower bounds.

Change-Id: Ic7929ac8747f463cd579649fd40925e9fb4365d0
Reviewed-on: http://review.couchbase.org/44877
Reviewed-by: Simon Baslé <simon@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-on: http://review.couchbase.org/44915
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Tested-by: Simon Baslé <simon@couchbase.com>