Correct KeyValueHandlerTests on master
Motivation
----------
Tests were failing after a modification ported from release10 to master.
This is due to the fact that on release10, JVMCBC-84 has not been ported
and thus KeyValueHandler still copy() the responseBuffer.

On master, said buffer is retain()ed, leading to a refCnt of 1 on the
response buffer.

Modifications
-------------
Fixed the assertions, content.refCnt() is expected to still be 1 since
it is the response buffer, retained in KeyValueHandler.decodeResponse.

Result
------
Unit test KeyValueHandlerTest now pass again on master (as it is passing
on release10).

Change-Id: I6764fa5581a57ed06e4ee18d916786a702cc18c9
Reviewed-on: http://review.couchbase.org/44986
Tested-by: Simon Baslé <simon@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>