Properly release copied content for counter.
Motivation
----------
In the backported release10 branch, the content is still copied. The counter
method still frees the original content, which leads to a illegal refcnt
exception, because it has been freed twice.

Modifications
-------------
Instead, the copied content should be freed since it is never passed up the
stack.

Result
------
No double freeing of the content when using the counter operation.

Change-Id: I041454ae254632c3c548909810686d7691d7f870
Reviewed-on: http://review.couchbase.org/45019
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>