JVMCBC-119: utility class for common ByteBuf ops
Motivation
----------
We often need to perform the same operations on ByteBufs, like release.
This class allows to centralize such common operations, especially in
the context of applying them in Rx streams.

Modifications
-------------
New Buffers utility class, with an Action1 implementation that allows
to release ByteBufs in a Rx stream.

Result
------
Less overhead when applying common operations on ByteBuf in some cases.

Change-Id: I250bab2a5ee5c82cd6e92438118595507da1c581
Reviewed-on: http://review.couchbase.org/45066
Tested-by: Simon Baslé <simon@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>