JVMCBC-121: Do not expose the list of partitions outside of the config.
Motivation
----------
The previous code did expose the List<Partition> to the whole core, but this makes
it very hard to change the underlying implementation.

Modifications
-------------
More getter methods have been exposed which provide the needed functionality, instead
of exposing the list directly. This allows to change the implementation of storage
later.

Result
------
Better encapsulation of the underlying data structures.

Change-Id: I805a5d144563bf3cbc631a99e942a134ef92f0c7
Reviewed-on: http://review.couchbase.org/45133
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>