Extract PartitionInfo in its own class.
Motivation
----------
The PartitionInfo, originally part of the CouchbaseBucketConfig, needs to be extracted
out so that proper unit tests can be added. It is large enough to stand alone.

Modifications
-------------
The class has been refactored out into the CouchbasePartitionInfo and along the way, the
forward map has been optimized away (only needed to check if tainted or not). In addition
one List is stored as an array since it never changes.

Result
------
Better testability and a unnecessary Lists removed.

Change-Id: I7415182cbc2e0f117e9db2edf02834d26a6f75a4
Reviewed-on: http://review.couchbase.org/45143
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>