JVMCBC-120: Prevent request if feature disabled
Motivation
----------
For some types of requests, if the corresponding feature is disabled in
Environment, the core has no handler and won't produce a Response on
the Observable. This can lead to hanging test/code.

Modifications
-------------
Such requests (for now Query and DCP) are identified and skipped in the
RequestHandler by erroring their Observable with an
UnsupportedOperationException.

Added tests to RequestHandlerTest to check that.

Result
------
Even if client of core forgets to check the environment for feature
activation, sending these Requests won't wait forever for a Response.

Change-Id: I45996d6b9fafab01bb22138bef19a4b1c5d76a50
Reviewed-on: http://review.couchbase.org/45225
Tested-by: Simon Baslé <simon@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>