JVMCBC-122: Release response content in handlers when removed.
Motivation
----------
When the view or query handler are removed for some reason (failover,
rebalance out, disconnect), the response content needs to be freed
since it is also allocated out of the pool.

Modifications
-------------
During handler cleanup, also release() the response content back to
the pool.

Result
------
No more held back bytebufs when a view or query handler  is removed
for some reason.

Change-Id: Ic30b498e0b7c133c95a95cf2cb1091baa62e0ac9
Reviewed-on: http://review.couchbase.org/45308
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>