JVMCBC-123: Release observe content properly.
Motivation
----------
When observe is used, the content needs to be properly freed. Actually,
the content is not needed at all, so it can be freed right in the key/value
handler.

Modifications
-------------
Free the content in the key/value handler after use for observe, but keep
an additional check in Observe just to be double safe on not leaking the
content.

Result
------
No leaks during observe poll cycles.

Change-Id: Id672f30f8a25dc2f955f1c4b225e87d36616b2ec
Reviewed-on: http://review.couchbase.org/45312
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>