Preserve environment invariant for feature tests
Motivation
----------
Feature checks are sensitive for system properties. And these systems
properties are the only way to adjust environment in cluster-dependent
tests as DCPMessageTest, which are not enabled by default, because the
server might lack of feature support.

But in case of tests, which verify the actual feature checking, should
work in the same way whether or not the system properties set.

Modification
------------
Reset system properties enabling Query and DCP features before executing
the tests, and restore on exit.

Result
------
Feature checking tests more robust

Change-Id: I97fe3f2a4c009323d9a52314fd502dc244281117
Reviewed-on: http://review.couchbase.org/45346
Tested-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>