JVMCBC-127: Send Opaque value for Key/value handler again.
Motivation
----------
Spymemcached did always fill the "opaque" field of the binary
protocol with an opaque value to make sure the returned response
matches up with the request as a sanity check.

Modifications
-------------
A static opaque value has been added that is now passed with every
binary request. On response, it is matched up with the request
to make sure all is well and if not, an exception is thrown. This
provides another line of defense for a sanity check.

Result
------
Better sanity checks on response parsing. It also helps when connecting
req/res in tcpdump and similar tools.

Change-Id: Ie4b60a2ca5bffb6ad80ab854388d4fab4f5b10e1
Reviewed-on: http://review.couchbase.org/45465
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>