JVMCBC-129: Allow reconnect delay to be configured.
Motivation
----------
Since we now have the Delay in the library, we can provide an easy
way to tune the reconnect delay when a socket is dropped.

Modifications
-------------
The changeset allows the reconnect delay to be configured from the
environment. The defaults are set similar to before, an exponential
delay with a ceiling of 4 seconds. The lower boundary has been lowered
a little from 128ms to 32ms.

Result
------
Completely configurable reconnect delay from the environment.

Change-Id: Ia288cdee7e8db5384e924edcb55acea2b48464fe
Reviewed-on: http://review.couchbase.org/45568
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>