JVMCBC-134: Add maximum request lifetime.
Motivation
----------
Without a maximum request lifetime, it is possible that requests
are circling around forever in the core.

Modifications
-------------
Adding a maximum request lifetime provides the "hard limit" for
requests that are highly likely to be removed anyways and prevents
them from filling up the request buffer endlessly.

Note that it needs to always be set higher than the maximum timeout
used in the application itself, otherwise it will prematurely
cancel valid request lifecycles.

This timeout is currently honored in the "best effort" strategy
and can be avoided by implementing a custom strategy.

Result
------
A hard limit when requests will be cleaned out of the system to
avoid having them circle around forever and filling up the request
buffer.

Change-Id: I6e7d8109fbec933df58fe8add41faa1cad3cedc6
Reviewed-on: http://review.couchbase.org/45923
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>