JVMCBC-135: Allow retryDelay to be configured.
Motivation
----------
The retry delay for operations was fixed to 10ms, which is not
flexible and maybe even too high. We can do better with our new
Delay object.

Modifications
-------------
This changeset makes the retryDelay configurable through the
environment and sets it to something more economic. By default,
it will exponentially increase from 100 micros to 100 millis instead
of the old default with a fixed delay of 10 millis.

Result
------
Configurable retry delay with more flexibility.

Change-Id: Ie743a4b2923a27f785928185748cd619456cdc2c
Reviewed-on: http://review.couchbase.org/45924
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>