JVMCBC-139: Add a generic EventBus and add events.
Motivation
----------
A generic EventBus is needed to publish different types of state
events to consumers through a "publish/subscribe" type interface.

This commit lays the foundation, adds the event bus itself and
defines some event types. Open/Close bucket and connect/disconnect
node events are also already emitted.

Modifications
-------------
This implementation adds a event bus based on a Rx subject and makes
sure to drop events if backpressure is not supported.

The CouchbaseNode pushes node connect/disconnect events and the
configuration manager open/close bucket events. Note that a
custom event bus can be passed in to the environment, but will
not be needed in general (maybe other than for testing).

Result
------
Basic foundation for more type of events in the future, like
errors or metrics.

Change-Id: I8eafb9d31e140b129c4766d8e18c6c6af2459df5
Reviewed-on: http://review.couchbase.org/45986
Reviewed-by: Simon Baslé <simon@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>