JVMCBC-66: also rewrite dep imports in original src
In sourceJar, since now we include shadowed dependencies sources, we
need to include rewritten original sources so that they import the
relocated classes and can be debugged.

The shadowSource task now also copies core sources and rewrite them to
reflect change in shadowed package names.

Only the result of shadowSource is included in the sourcesJar task.

Change-Id: I2c2996f5f6e4b6144277b3179077674529168847
Reviewed-on: http://review.couchbase.org/46227
Tested-by: Simon Baslé <simon@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>