JVMCBC-141: Detect and ignore string in processor
Motivation
----------
ClosingPositionBufProcessor currently doesn't account for strings in
JSON, which could throw its detection off if the string contains an
uneven number of opening and closing characters.

Modifications
-------------
The processor now has an option to detect and ignore strings after the
first detected section opening. Added tests for this behavior.

Results
-------
All test pass and the processor can now ignore strings in the processed
content (provided it is JSON data, not starting inside a JSON string).

Change-Id: Ic374899947aae0da3537cc163b030101321f010a
Reviewed-on: http://review.couchbase.org/46434
Tested-by: Simon Baslé <simon@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>