JVMCBC-142: Ignore escaped quotes in clientID
Motivation
----------
When parsing a clientContextId, the id could contain escaped quotes.
Currently the parser doesn't manage that and it will result in bad
parsing.

Modifications
-------------
Parsing of clientContextId now uses the method that ignores escaped "
when searching for the closing quote.

Results
-------
ClientContextId containing an escaped quote are now correctly parsed.

Change-Id: I8ec580ac49d40d30176bd9cd3ec24da06bf778b9
Reviewed-on: http://review.couchbase.org/46435
Tested-by: Simon Baslé <simon@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>