JVMCBC-146: Fail the Observe sequence when replica is not configured on the bucket.
Motivation
----------
When the user provides a persistence requirement which is not doable in the first place
because it is larger than the number of replicas configured on the bucket, it makes
sense to fail this observable quickly (since it will never be completed anyways).

Modification
------------
When the number of replicas is exctracted from the bucket config, a check is now
performed to make sure the given replica requirement is in range with what is
configured on the bucket (<=).

If it is not doable, fail with an exception.

Result
------
Quicker and semantically correct feedback to a user in this case.

Change-Id: I9c844abe2c66c39004130f210dacf5bfc4591785
Reviewed-on: http://review.couchbase.org/46590
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>