JVMCBC-147: Fail Observe when document is modified concurrently.
Motivation
----------
It is possible that while one request is still polling anotherone
can interfer with the previous one, change the CAS and lead to a
observe "Modified" response.

To give this information to the user, it needs to be properly
caught and propagated.

Modifications
-------------
The MODIFIED response from observe is explicitly caught and the
observe observable is failed. The user can catch it and either
retry or propagate the error. In any case it needs to fail because
the persistence requirement for this particular ID is not
satisfiable anymore.

Result
------
Proper error raised when the concurrent modification error is
happening.

Change-Id: I6ab469226ca85429fd444abcb925e09a93d010aa
Reviewed-on: http://review.couchbase.org/46591
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>