JVMCBC-149: Allow to fail fast during observe poll cycles
Motivation
----------
When a fail fast strategy is used, observe poll cycles also need to
return instantly and not take longer than it is needed to decide
if the desired state can be reached or not.

Modifications
-------------
The retry strategy now also has a second method where one can decide
if observe polls should be retried or not, and the default strategies
are implemented properly.

If an error happens on one of the polling observables, it either gets
swallowed and the polling goes on until a good state is reached or
the whole thing fails immediately.

Result
------
Proper fail fast characteristics when durability requirements are
used.

Change-Id: I1509d3c5bf60b2033bc4062df4459d72a8e8bc7c
Reviewed-on: http://review.couchbase.org/46594
Reviewed-by: Simon Baslé <simon@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>