JVMCBC-138: Fix incorrect logging with IDLE states.
Motivation
----------
When the IDLE state was added, the code which checks if the
state is CONNECTED on the Node level has been updated, but
did not respect the situation where only one IDLE service
is part of the node.

This results in prematurely promoting a CONNECTED state,
which can quickly go back to DISCONNECTED if the next
service comes along that is not an on demand one.

Modifications
-------------
If the node only has on demand services by now, do not
count an IDLE state as CONNECTED, but rather expose
it as an explicit state as well.

Result
------
No misleading log messages on bootstrap that a node
"bounces" between connected and disconnected states.

Change-Id: I8a55e26f858f1bd80c1bc2ea1d936193a569d91c
Reviewed-on: http://review.couchbase.org/46874
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>