JVMCBC-136: Automatically release not subscribed ByteBufs on View and Query
Motivation
----------
When View or N1Ql queries are run and then not all subjects are properly subscribed
to (which can easily happen if for example an error happens and one only listens to
rows) then the code leaks ByteBufs.

Modifications
-------------
The changeset borrows a unicast auto release subject from RxJava which handles
autoreleasing after some period, which is set to the maxRequestLifetime of the
environment.

Result
------
Avoid leaking even if requested but not subscribed.

Change-Id: I66f13a565d59898ec03ecbefd0a9c8a9aead3622
Reviewed-on: http://review.couchbase.org/46942
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>