JVMCBC-151: fix broken query keepAlive
Motivation
----------
In QueryHandler, keepalive is broken: it is marked as finishedDecoding
too soon, not taking into account that the first response could not be
a LastHttpComponent. This can break the handling of query requests.

Modifications
-------------
Only finish the decoding and issue the response for keepalive when we
received a LastHttpComponent.

Results
-------
Keepalive doesn't interfere with query handling anymore.

Change-Id: I445f86596e629e1d0874f4606978e50019b09666
Reviewed-on: http://review.couchbase.org/46943
Tested-by: Simon Baslé <simon@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>