JVMCBC-137: Determine based on the config if a service is enabled or not.
Motivation
----------
The code properly needs to determine if a request can be served at all
or not. Previously it only did check on the environment flags, but since
the service can be announced through the config a more elaborate
logic needs to be put in place.

Modifications
-------------
The configuration now exposes a "serviceEnabled" method which returns
true/false. The data is cached on creation to make sure the data does
not need to be computed over and over.

It will return true if at least one node in the config is able to
serve it, so it can change at runtime.

Result
------
More accurate dynamic rejecting of requests which are not servable
based on the current config.

Change-Id: Iea00031ae45ee1ec9aba2da67f60601722649688
Reviewed-on: http://review.couchbase.org/47276
Reviewed-by: Simon Baslé <simon@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>