JVMCBC-156: Manage the Node service states through the state zipper.
Motivation
----------
In 1.1.0 the state zipper was added to reliably managed a list of
changing stateful components. We can reuse the same functionality
for the node as well.

Modifications
-------------
Moved the registration to the zipper while retaining the computation
logic for the actual log messages and events.

Result
------
Centralized management for states.

Change-Id: I36bdc9bc8c80b09cb06e38a604f37d2443c739b3
Reviewed-on: http://review.couchbase.org/47405
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>