JVMCBC-154 - Avoid leaking content on not needed config buffers.
Motivation
----------
The carrier loader and refresher both use the bucket config result, but it could
be that the code unsubscribes before the full result is consumed. This is a valid
condition, but code needs to be in place to free all the buffers nonetheless.

Modifications
-------------
Since the generic "release if unsubscribed" behaviour can be reused in lots of places,
a helper method has been introduced to the Buffers class which transparently takes
care.

The loader and refresher have been adapted to account for the automatic release.

Result
------
Even when unsubscribed on config loading, the buffers are properly released. Also
the helper class can be used in various other places as well.

Change-Id: I95c3edcd8b86e36e7328adb76c8585d5b4c992ee
Reviewed-on: http://review.couchbase.org/47855
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>