JVMCBC-162: Do not flush if not active.
Motivation
----------
Much like on write(), flush() needs to check if the channel is
active and only then send the flush signal.

Modifications
-------------
Check if the channel is active and only then flush. It makes no
sense to flush if the channel is not active, error handling
facilities inside of the handler perform the necessary
cleanup tasks.

Result
------
No flushing when the channel is not active.

Change-Id: I8df2f806701f51d7332a809cc24f72e2b663e41e
Reviewed-on: http://review.couchbase.org/47859
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>