JVMCBC-163: Avoid eager evaluation of config fallback.
Motivation
----------
Even when onErrorResumeNext is done once a failure happens, passing
it in right away will make the JVM eagerly evaluate the call
and trigger the network request.

Modifications
-------------
Use Observable.defer to correctly defer the hot observable call and
avoid a request which is never consumed and potentially leaks the
content.

Result
------
No eager evaluation and no leak.

Change-Id: If2fd84f3a4b724955a444a854c40e0ec384e8f62
Reviewed-on: http://review.couchbase.org/47861
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>