JVMCBC-164: Hide IOExceptions on write failures.
Motivation
----------
The write callback error log is designed to surface coding issues
that are not logged by the netty event loop. When a channel is
closed forcefully (hard failover on the server), IOExceptions
including broken pipes are logged.

This information is redundant and confuses the user potentially,
since the SDK is expected to handle that.

Modifications
-------------
The IOExceptions are ignored and not logged, because the SDK
already indicates to the user that a node has been disconnected
in this case. Also, a broken pipe can happen as a race condition
and the handler is expected to clean everything up nicely.

Result
------
Less cognitive load for the user and support to grasp during a
hard channel close.

Change-Id: I1b4ab47845bf218c4880ed7b4bb1e99a2a9c97ab
Reviewed-on: http://review.couchbase.org/47862
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>