extract dumping of env parameters for JCBC-724
Motivation
----------
In the java-client, subclass of DefaultCoreEnvironment has several
specific parameters, that should also be dumped in the toString. This
change is needed in order for this subclass to use core's dumping of
parameters

Modifications
-------------
A new protected method that dumps core environment parameters into a
StringBuilder has been put in place, and toString relies on it.

Results
-------
Subclasses can now dump the full range of environment parameters in a
StringBuilder.

Change-Id: I80753f7598e5affc0adc11c3ec89207260c1ba0c
Reviewed-on: http://review.couchbase.org/48005
Tested-by: Simon Baslé <simon@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>